new shit

+++ include/pn532.h
@@ -61,7 +61,7 @@ public:
   std::vector<uint8_t> get_firmware_version();
   std::vector<uint8_t> read_passive_target(int baud=PN532_MIFARE_ISO14443A);
   std::vector<uint8_t> ultralight_read_page(int page);
-  int ultralight_write_page(std::vector<uint8_t> dat, int page);
+  uint8_t ultralight_write_page(std::vector<uint8_t> dat, int page);
 
 private:
   int DEBUG_LVL;
